+++ src/main.rs
@@ -2,23 +2,22 @@ mod backend;
 mod config;
 
 use backend::Backend;
+use clap::Parser;
 use log::debug;
 use simplelog::{Config, LevelFilter, WriteLogger};
 use std::fs::File;
-use structopt::{clap, StructOpt};
 use tower_lsp::{LspService, Server};
 
 // -------------------------------------------------------------------------------------------------
 // Opt
 // -------------------------------------------------------------------------------------------------
 
-#[derive(Debug, StructOpt)]
-#[structopt(name = "svls")]
-#[structopt(long_version(option_env!("LONG_VERSION").unwrap_or(env!("CARGO_PKG_VERSION"))))]
-#[structopt(setting(clap::AppSettings::ColoredHelp))]
+#[derive(Debug, Parser)]
+#[clap(name = "svls")]
+#[clap(long_version(option_env!("LONG_VERSION").unwrap_or(env!("CARGO_PKG_VERSION"))))]
 pub struct Opt {
     /// Debug mode
-    #[structopt(short = "d", long = "debug")]
+    #[clap(short = 'd', long = "debug")]
     pub debug: bool,
 }
 
@@ -28,7 +27,7 @@ pub struct Opt {
 
 #[tokio::main(flavor = "current_thread")]
 async fn main() {
-    let opt = Opt::from_args();
+    let opt: Opt = Parser::parse();
 
     if opt.debug {
         let _ = WriteLogger::init(